prevent compiler warnings for VC-WIN64A

+++ crypto/crmf/crmf_pbm.c
@@ -51,11 +51,11 @@ CRMF_PBMPARAMETER *CRMF_pbmp_new(size_t slen, int owfnid,
         error = CRMF_R_MALLOC_FAILURE;
         goto err;
     }
-    if (RAND_bytes(salt, slen) <= 0) {
+    if (RAND_bytes(salt, (int)slen) <= 0) {
         error = CRMF_R_FAILURE_OBTAINING_RANDOM;
         goto err;
     }
-    if (!(ASN1_OCTET_STRING_set(pbm->salt, salt, slen)))
+    if (!(ASN1_OCTET_STRING_set(pbm->salt, salt, (int)slen)))
         goto err;
 
     /*
